% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGini.R
\name{scGini}
\alias{scGini}
\title{Calculate corrected Gin index}
\usage{
scGini(mat, ncore = 1)
}
\arguments{
\item{mat}{Single-cell counts matrix with genes as rows and cells as columns.}

\item{ncore}{Number of cores for parallel computing, default is 1.}
}
\value{
A data.frame with estimated parameters and directly calculated Gini index (Gini)
and corrected Gini index (cGini).
}
\description{
This function calculates corrected Gini index from single cell RNA-seq counts data
without any preprocessings.
}
