% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGiniSmooth.R
\name{scGiniSmooth}
\alias{scGiniSmooth}
\title{Neighborhood Smoothing of the scGini}
\usage{
scGiniSmooth(mat, scgini, nGenes = 1000)
}
\arguments{
\item{scgini}{Corrected Gini index, results from \code{scGini}}

\item{nGenes}{Number of genes used for neighborhood smoothing}

\item{dat}{Single-cell counts matrix with genes as rows and cells as columns.}
}
\value{
A vector with smoothed scGini index is returned.
}
\description{
This functions implements a neighborhood difussion map to improve the scGini prediction
by utilizing information from neighboring cells.
}
